feat: implement findLongestCommonSubarray function and add comprehensive tests; remove unused diffArray and getBiggerClusterIndex tests
